chore: build common barchart
